Refresh comments on tab activation, if necessary
If the page is loaded while some tab other than the
comments tab is active (comments load invisibly),
the iframe will detect its height as `0` and fail to
resize properly.

When activating the comments tab, check the
height property of the embed, and refresh if
it's 0 or hasn't been set.